update test feedback changes

+++ engineering/models.py
@@ -107,6 +107,7 @@ class TrainingAgenda(TimeStampedModel):
     def __str__(self):
         return str(self.project.id)
 
+
 class Cycle(models.Model):
     is_current = models.BooleanField(_('Is current'), default=True)
     start_date = models.DateField(_('Start Date'))
@@ -118,8 +119,9 @@ class Cycle(models.Model):
         else:
             return f"Cycle ({self.start_date} to {self.end_date})"
 
+
 class EngineerPoint(TimeStampedModel):
-    points = models.IntegerField(_('points'), default=0)
+    points = models.DecimalField(_('points'), default=0.0, max_digits=10, decimal_places=2)
     is_active = models.BooleanField(_('Is active'), default=True)
     cycle = models.ForeignKey(
         Cycle, on_delete=models.PROTECT,
